requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.26
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
